requirements.txt
Flask
cryptography       # For Fernet, AES
Pillow             # For image encoding
pycryptodome       # For AES, RSA